[workitem:75] - Adding a simple implementation of "complete" from Readline/BASH.  This only does the displaying of common prefixes.
I have found that this breaks certain types of tab completion in PowerTab.  Most notably where the the text and values of options are different and the common prefix is also one of the options.

Example:  Stop-Process -Id powershell<tab>  ## you will never get an ID if you have PowerShell and ISE running, or more than one of either running